requirements.txt
PyPDF2
openai==0.27.7
pinecone-client==3.1.0
python-dotenv
numpy
flask
Werkzeug